orm/requirements.txt
asgiref==3.10.0
Django==5.2.7
django-debug-toolbar==6.0.0
django-extensions==4.1
sqlparse==0.5.3
